Allow subclasses of classes implementing Redis::Objects to automatically pick up objects from their superclasses.
This is a two-pronged approach.  First, when we call one of the declarative class methods that defines the
accessors for the object, we hardcode the class reference for access @redis and @redis_objects.  The second
part is to update the field_key methods in the core Redis::Objects module so that they explicitly look up the
class to reference via first_ancestor_with.  This ensures that the key name is consistent with where the object
was declared.

Tests cover the base inheritance case where no further redis objects are declared (which should make it compatible
with ActiveRecord STI I'm hoping, but that's untested), as well as the case where an object is overriden by a
subclass, and a new object is declared.