Add delete! method to Redis::Objects
As a first attempt to add a feature to this great project, I added
a method to Redis::Objects that allows an instance to delete itself.

Instead of a user having to go through each object themselves and
deleting each piece, this deletes the whole object for you.

Also added a test for the method.