Update warning regarding expiration settings
Based on https://github.com/nateware/redis-objects/pull/213 and the comments on https://github.com/nateware/redis-objects/commit/4854553f451c1b983ce620e205c45acea1fefc5b, the current README incorrectly describes an issue with the `expiration` setting when it is actually meant to be describing an issue with the `expireat` setting. This updates the documentation to correctly describe the issue.